Fix adopt-redthread importer module size
